+++ src/main/java/com/cookeep/cookeep/domain/user/application/AuthService.java
@@ -11,8 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.cookeep.cookeep.api.dto.request.TokenRefreshRequestDTO;
 import com.cookeep.cookeep.api.dto.response.KakaoLoginResponseDTO;
 import com.cookeep.cookeep.api.dto.response.TokenRefreshResponseDTO;
-import com.cookeep.cookeep.config.JwtTokenProvider;
-import com.cookeep.cookeep.domain.onboarding.entity.UserOnboarding;
+import com.cookeep.cookeep.security.JwtTokenProvider;
 import com.cookeep.cookeep.domain.user.dao.UserAuthRepository;
 import com.cookeep.cookeep.domain.user.dao.UserRepository;
 import com.cookeep.cookeep.domain.user.dao.UserSessionRepository;
