+++ src/main/java/com/cookeep/cookeep/security/UserPrincipal.java
@@ -1,4 +1,4 @@
-package com.cookeep.cookeep.config;
+package com.cookeep.cookeep.security;
 
 // 현재 로그인한 사용자의 userId 정보를 담는 DTO
 public record UserPrincipal(Long userId) {
